+++ DiploGM/manager.py
@@ -3,6 +3,8 @@ import time
 import os
 from typing import Optional
 
+from discord import Member
+
 from DiploGM.utils.singleton import SingletonMeta
 from DiploGM.adjudicator.adjudicator import make_adjudicator
 from DiploGM.adjudicator.mapper import Mapper
@@ -12,6 +14,7 @@ from DiploGM.models.board import Board
 from DiploGM.db import database
 from DiploGM.models.player import Player
 from DiploGM.models.spec_request import SpecRequest
+from DiploGM.utils.sanitise import simple_player_name
 
 logger = logging.getLogger(__name__)
 
@@ -313,3 +316,10 @@ class Manager(metaclass=SingletonMeta):
         message = f"Reloaded board for phase {loaded_board.turn.get_indexed_name()}"
         file, file_name = mapper.draw_current_map()
         return {"message": message, "file": file, "file_name": file_name}
+
+    def get_member_player_object(self, member: Member) -> Player | None:
+            for role in member.roles:
+                for player in self.get_board(member.guild.id).players:
+                    if simple_player_name(player.name) == simple_player_name(role.name):
+                        return player
+            return None